+++ src/renderer/src/components/pet/PetOverlay.tsx
@@ -4,6 +4,7 @@ import { usePetUrl } from './usePetUrl'
 import type { DetectedSpriteCacheEntry } from './pet-blob-cache'
 import type { CustomPet } from '../../../../shared/pet-types'
 import { useAppStore } from '../../store'
+import { getAgentStatusEpochNow } from '@/lib/agent-status-epoch-clock'
 import { AGENT_STATUS_STALE_AFTER_MS } from '../../../../shared/agent-status-types'
 import {
   selectPetAnimationName,
@@ -23,14 +24,10 @@ function usePetAnimationName(
   const agentStatusByPaneKey = useAppStore((s) => s.agentStatusByPaneKey)
   const agentStatusEpoch = useAppStore((s) => s.agentStatusEpoch)
   const retainedAgentsByPaneKey = useAppStore((s) => s.retainedAgentsByPaneKey)
-  const [agentStatusNow, setAgentStatusNow] = useState(() => Date.now())
-  useEffect(() => {
-    setAgentStatusNow(Date.now())
-  }, [agentStatusEpoch])
-
   // Re-render when the freshness scheduler ticks so stale live states stop
-  // driving pet animations even if no other store value changes.
-  void agentStatusEpoch
+  // driving pet animations even if no other store value changes, and read the
+  // boundary clock in render so the stale frame never paints.
+  const agentStatusNow = getAgentStatusEpochNow(agentStatusEpoch)
 
   return selectPetAnimationName({
     entries: Object.values(agentStatusByPaneKey),
