+++ src/renderer/src/components/sidebar/use-visible-workspace-kanban-worktree-ids.ts
@@ -1,5 +1,6 @@
-import { useEffect, useMemo, useState } from 'react'
+import { useMemo } from 'react'
 import { useAppStore } from '@/store'
+import { getAgentStatusEpochNow } from '@/lib/agent-status-epoch-clock'
 import type { Repo } from '../../../../shared/repo-types'
 import type { Worktree } from '../../../../shared/worktree/types'
 import { computeVisibleWorktrees } from './visible-worktrees'
@@ -51,10 +52,9 @@ export function useVisibleWorkspaceKanbanWorktreeIds({
     !showSleepingWorkspaces ? s.browserTabsByWorktree : null
   )
   const agentStatusEpoch = useAppStore((s) => (!showSleepingWorkspaces ? s.agentStatusEpoch : 0))
-  const [agentStatusNow, setAgentStatusNow] = useState(() => Date.now())
-  useEffect(() => {
-    setAgentStatusNow(Date.now())
-  }, [agentStatusEpoch])
+  // Why: skip the clock entirely when the epoch is the opt-out sentinel, so a
+  // sleeping-workspaces board cannot evict the sample the live boards share.
+  const agentStatusNow = showSleepingWorkspaces ? 0 : getAgentStatusEpochNow(agentStatusEpoch)
   // Why snapshot on the epoch: the always-mounted drawer must not scan every
   // agent on unrelated store writes; membership changes advance this tick.
   const worktreeIdsWithLiveAgent = useMemo(() => {
